requirements.txt
fastapi[standard]
uvicorn[standard]
gunicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
celery[redis]
redis
langgraph
langchain-openai
langchain-anthropic
langchain-ollama
openai
boto3==1.42.42
httpx
pydantic-settings
python-dotenv
slowapi
python-multipart
streamlit
requests
gtts